Fix wrong luminosity evolution correction
The expression for the 2MRS luminosity evolution correction taken from
Branchini et al. (2012) contained a typo. This led to an underestimation
of the selection function towards the reconstruction boundary, which in
turn created a spurious outflow in the reconstruction. Now the
luminosity evolution correction and, for consistency, also the
k-correction from Bell et al. (2003) are used, instead. They yield a
selection function that correctly matches the observed redshift
distribution.