Correct for small spurious mean density contrast
The RSD correction can introduce a very small spurious mean density
contrast in the reconstruction volume.  In combination with the
Fourier-Bessel boundary conditions, which set the density contrast to
zero outside of this volume, this can then create a small spurious in-
or outflow. This is now accounted for by adding a monopole contribution
after performing the RSD correction, which forces the r-space mean
density contrast to be zero.