Add cosmic variance to tensor-smoothing weighting
Following Dekel et al. (1990), a contribution due to the cosmic variance
of the velocity field is added to the quadratic observed velocity
errors in the weighting of the tensor-smoothing. To compute this cosmic
variance within the Gaussian tensor-smoothing window of a certain width,
a method is added to NormalizedPowerSpectrum, which computes
the variance of the density contrast weighted with some k-dependent
kernel.