Slightly change definition of group redshifts
Previously group redshifts were defined by averaging galaxy redshifts in
the CMB frame, and then transforming the result to the Heliocentric (HC) and
Local Group (LG) frames. Now the averaging is performed in the
originally observed HC frame, instead. In practice, the difference is
negligible. It only changes a couple of the HC group redshifts (used
as input for the reconstruction) by 1 km/s.